allow to insert strings as supercollider symbols
This commit changes the convert_to_sc function to format strings
starting with a single / (slash) in such a way that they are
evaluated as symbols by supercollider.
If the string starts with double // the first / will be omitted,
this allows the user to send strings starting with / to the
supercollider language.